Create policy for Friendships destroy action
- A friendship can be rejected by the receiver or terminated by any
participant, in both cases the record is deleted

+++ app/controllers/friendships_controller.rb
@@ -22,6 +22,8 @@ class FriendshipsController < ApplicationController
 
   def destroy
     friendship = Friendship.find(params[:id])
+    authorize friendship
+
     friendship.destroy
 
     redirect_back(fallback_location: root_path)
